pickle-ai/pyproject.toml
[project]
name = "pickle-ai"
version = "0.1.0"
description = "AI Persona companion for IRL Twitch streamers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "twitchio>=2.6.0,<3.0.0",
    "google-genai>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "deepgram-sdk>=3.9.0,<4.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "tokenc>=0.1.0",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/persona", "config"]

[project.scripts]
pickle-ai = "src.persona.main:run"
